add catagoerie + securisation des routes

+++ src/Controller/ArticleController.php
@@ -17,9 +17,13 @@ class ArticleController extends AbstractController
     #[Route('/', name: 'article_index', methods: ['GET'])]
     public function index(ArticleRepository $articleRepository): Response
     {
-        return $this->render('article/index.html.twig', [
-            'articles' => $articleRepository->findAll(),
-        ]);
+        if ($this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            return $this->render('article/index.html.twig', [
+                'articles' => $articleRepository->findAll(),
+            ]);
+        }else{
+            return $this->redirectToRoute('article_user');
+        }
     }
 
     #[Route('/user', name: 'article_user', methods: ['GET'])]
@@ -33,21 +37,27 @@ class ArticleController extends AbstractController
     #[Route('/new', name: 'article_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
-        $article = new Article();
-        $form = $this->createForm(ArticleType::class, $article);
-        $form->handleRequest($request);
+        if ($this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            
+            $article = new Article();
+            $form = $this->createForm(ArticleType::class, $article);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager->persist($article);
+                $entityManager->flush();
+    
+                return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+            }
+    
+            return $this->renderForm('article/new.html.twig', [
+                'article' => $article,
+                'form' => $form,
+            ]);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($article);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+        }else{
+            return $this->redirectToRoute('article_user');
         }
-
-        return $this->renderForm('article/new.html.twig', [
-            'article' => $article,
-            'form' => $form,
-        ]);
     }
 
     #[Route('/{id}', name: 'article_show', methods: ['GET'])]
@@ -61,29 +71,38 @@ class ArticleController extends AbstractController
     #[Route('/{id}/edit', name: 'article_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(ArticleType::class, $article);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+        if ($this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $form = $this->createForm(ArticleType::class, $article);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager->flush();
+    
+                return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+            }
+    
+            return $this->renderForm('article/edit.html.twig', [
+                'article' => $article,
+                'form' => $form,
+            ]);
+        }else{
+            return $this->redirectToRoute('article_user');
         }
-
-        return $this->renderForm('article/edit.html.twig', [
-            'article' => $article,
-            'form' => $form,
-        ]);
     }
 
     #[Route('/{id}', name: 'article_delete', methods: ['POST'])]
     public function delete(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($article);
-            $entityManager->flush();
+        if ($this->container->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($article);
+                $entityManager->flush();
+            }
+    
+            return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+        }else{
+            return $this->redirectToRoute('article_user');
         }
-
-        return $this->redirectToRoute('article_index', [], Response::HTTP_SEE_OTHER);
+        
     }
 }
